added labels and validator in the new file
